refactor: model => modal
