+++ src/components/platform/community-modals/AddNewGroup.tsx
@@ -1,9 +1,9 @@
 import { useForm } from "react-hook-form";
-import { TextControl } from "./model/controls/TextControl";
-import Modal from "./model/Model";
-import { Divider, Error } from "./model/Content";
+import { TextControl } from "./modal/controls/TextControl";
+import Modal from "./modal/Modal";
+import { Divider, Error } from "./modal/Content";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { SelectControl } from "./model/controls/SelectControl";
+import { SelectControl } from "./modal/controls/SelectControl";
 import { useEffect, useState } from "react";
 import * as yup from 'yup';
 import { fetchTopics } from "./common.schema";
