+++ src/components/platform/community-modals/add-new-event.schema.ts
