+++ src/components/platform/community-modals/add-new-group.schema.ts
