+++ src/components/platform/community-modals/add-new-page.schema.ts
