+++ src/components/platform/community-modals/common.schema.ts
