+++ src/components/platform/community-modals/modal/Content.tsx
