+++ src/components/platform/community-modals/modal/actions.module.scss
@@ -39,7 +39,7 @@
             cursor: pointer;
 
             &:hover {
-                scale: 1.2;
+                scale: 1.05;
             }
         }
