+++ src/components/platform/community-modals/modal/content.module.scss
