+++ src/components/platform/community-modals/modal/controls/TextControl.tsx
